+++ CHANGELOG.md
@@ -63,6 +63,7 @@ and disable fields when "Settings (Inventory): Create, edit and delete HRID hand
 * Edit in Linked Data Editor when source=MARC. Refs UIIN-2978.
 * Display shared instance immediately after the instance was shared. Refs UIIN-2969.
 * Disable Move to button if no items selected or no more holdings exists within an instance. Refs UIIN-2995.
+* Prevent callbacks from being triggered when clicking on the current tab, whether it's lookup mode or a segment. Fixes UIIN-2949.
 
 ## [11.0.4](https://github.com/folio-org/ui-inventory/tree/v11.0.4) (2024-04-30)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v11.0.3...v11.0.4)